+++ locales-loader.js
@@ -1,4 +1,4 @@
-const buildConfig = require('./server/config.app');
+const buildConfig = require('./config.app');
 
 function localesLoader(source) {
   if (buildConfig.TEST) {
