+++ package.json
@@ -15,9 +15,11 @@
     "test-cloudfront": "bash test/lambda/run-lambda-tests.sh",
     "start-dev": "NODE_ENV=development USE_WEBPACK_DEV_SERVER=true ALLOW_SIGNUP_PATIENT=false ALLOW_PATIENT_CHANGE_EMAIL=false I18N_ENABLED=true ALLOW_PATIENT_CHANGE_PASSWORD=false HIDE_DONATE=true HIDE_DEXCOM_BANNER=true HIDE_UPLOAD_LINK=true CAN_SEE_PWD_LOGIN=true BRANDING='diabeloop' METRICS_SERVICE='disabled' HELP_LINK='disabled' webpack-dev-server --cache --colors --progress --port 3001 --host 0.0.0.0",
     "prebuild": "rm -rf dist server/dist deploy",
-    "build": "npm run build-app && npm run build-config",
+    "build": "npm run build-app && npm run gen-lambda",
+    "gen-lambda": "NODE_ENV=production npm install && BLIP_VERSION=$(echo \"const c = require('./package.json'); console.log(c.version);\" | node -) node cloudfront-dist/cloudfront-gen-lambda.js",
+    "deploy-cf": "echo 'not implemented yet'",
     "prebuild-app": "echo 'Building app'",
-    "build-app": "NODE_ENV=production webpack",
+    "build-app": "NODE_ENV=production webpack && mv static-dist/index.html cloudfront-dist/pre-compiled-index.html",
     "build-config": "NODE_ENV=production node server/build-config.js",
     "prebuild-dev": "rm -rf dist server/dist",
     "build-dev": "NODE_ENV=development webpack --progress",
